fix: (BaseTile) Fix issues with base tile and update stories
